+++ CHANGELOG.md
@@ -37,3 +37,5 @@
 and `crypto_auth_verify` functions of the `NaCl` library (does not exists in TweetNaCl).
 - Added HMAC-SHA-512.
 - Added HMAC-SHA-512 unit tests.
+- Added some `TweetNaCl`'s tests.
+- Cleaned some code.